Support for weighted SAT problems
This patch extends MonadSAT with a method emitW
for weighted clauses and provides a new concrete
implementation of MonadSAT, Weighted.SAT, 
which can do weighted problems.

The previous SAT monad implementation is preserved
for solvers which do not support weighted SAT, as
e.g. miniSAT

This patch also moves the generation of the DIMACS
header to the Solver, as there is more that one
format convention for weighted SAT problems